+++ src/lib/router/header-route.ts
@@ -1,6 +1,6 @@
 export const headerRouteConfig = [
   { route: 'home', navName: 'Home' },
-  { route: 'timetable', navName: 'Timetable', innerTab: ['timetable', 'timetable/friend'] },
+  { route: 'timetable', navName: 'Timetable', innerTab: ['timetable', 'friend'] },
   { route: 'community', navName: 'Community' },
   { route: 'matching', navName: '1:1 Matching' },
 ]
