+++ src/lib/router/lazy-route.tsx
@@ -24,7 +24,6 @@ export const WriteReviewPage = withSuspense(lazy(() => import('@/pages/CourseRev
 export const MyPage = withSuspense(lazy(() => import('@/pages/MyPage')))
 export const SchedulePage = withSuspense(lazy(() => import('@/pages/SchedulePage')))
 export const TimetablePage = withSuspense(lazy(() => import('@/pages/TimetablePage')))
-export const FriendPage = withSuspense(lazy(() => import('@/pages/TimetablePage/FriendPage')))
-export const FriendTimetablePage = withSuspense(lazy(() => import('@/pages/TimetablePage/FriendTimetablePage')))
-export const MyTimetablePage = withSuspense(lazy(() => import('@/pages/TimetablePage/MyTimetablePage')))
+export const FriendPage = withSuspense(lazy(() => import('@/pages/FriendPage')))
+export const FriendTimetablePage = withSuspense(lazy(() => import('@/pages/FriendTimetablePage')))
 export const ClubDetailPage = withSuspense(lazy(() => import('@/pages/ClubPage/ClubDetailPage')))
